requirements.txt
azure-ai-formrecognizer==3.3.3
azure-core==1.30.1
openai==1.30.4
